Rafactor loggers into loggers.py
